added a log if the station that is tried to play is different to what is already playing on a different device
this is a limitation of owntone, it cant play two different tracks to two different devices at the same time.
I have solved this: firstly when the switch is switched on, the item is added to the queue as normal, but after that i get the queue. i then compare all the items in the queue to check for duplicates. if there are all duplicates, the same as before happens (device is activated and queue played) however, if one item in the queue is not the same as the rest, i get an array of all the items in the queue that are different to what is currently playing. I then loop through this array and call the api to remove each of these ids from the queue.
still todo for this feature:
in the getonhandler i need another check to see if what is currently playing matches what the config states, it is matches, the switch is on, if it doesn't, the switch is off